# Conflicts:
#	engine-master/engine-master-server/src/test/java/com/dtstack/engine/master/impl/ClusterServiceTest.java

+++ engine-master/engine-master-server/src/test/java/com/dtstack/engine/master/impl/ClusterServiceTest.java
@@ -195,7 +195,7 @@ public class ClusterServiceTest extends AbstractTest {
 
         //删除组件
         try {
-            componentService.delete(Lists.newArrayList(hdfsComponent.getId().intValue()));
+            componentService.delete(Lists.newArrayList(hdfsComponent.getId()));
         } catch (Exception e) {
             if (e instanceof RdosDefineException) {
                 RdosDefineException rdosDefineException = (RdosDefineException) e;
